mock_services/user/pyproject.toml
[project]
name = "user-service"
version = "0.1.0"
description = "A simple mock user service"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "httpx>=0.28.1",
    "opentelemetry-distro>=0.55b1",
    "opentelemetry-exporter-otlp>=1.34.1",
    "opentelemetry-instrumentation-asgi>=0.55b1",
    "opentelemetry-instrumentation-fastapi>=0.55b1",
    "opentelemetry-instrumentation-httpx>=0.55b1",
    "opentelemetry-instrumentation-logging>=0.55b1",
    "uvicorn>=0.34.3",
]
